Se ha traducido todo al castellano.

+++ bewomByte/src/main/java/es/bewom/warps/commands/CommandWarpDel.java
@@ -12,6 +12,8 @@ import org.spongepowered.api.util.command.spec.CommandExecutor;
 
 import com.google.common.base.Optional;
 
+import es.bewom.texts.TextMessages;
+import es.bewom.user.BewomUser;
 import es.bewom.warps.WarpManager;
 
 public class CommandWarpDel implements CommandExecutor {
@@ -21,12 +23,18 @@ public class CommandWarpDel implements CommandExecutor {
 			throws CommandException {
 		
 		if(!(src instanceof Player)) {
-			src.sendMessage(Texts.of("This command is for players only."));
+			src.sendMessage(TextMessages.NOT_CONSOLE_COMPATIBLE);
 			return CommandResult.empty();
 		}
 		
 		Player player = (Player) src;
-		String warpName = args.<String>getOne("name").get();
+		BewomUser user = BewomUser.getUser(player);
+		if(user.getPermissionLevel() < BewomUser.PERM_LEVEL_ADMIN) {
+			player.sendMessage(TextMessages.NO_PERMISSIONS);
+			return CommandResult.empty();
+		}
+		
+		String warpName = args.<String>getOne("nombre").get();
 		
 		Optional<String> error = WarpManager.deleteWarp(warpName);
